+++ src/config.h
@@ -188,6 +188,7 @@ typedef struct {
   string suppress_nrc;
   string suppress_wheel;
   string filter_paste;
+  int bracketed_paste_split;
   int suspbuf_max;
   int printable_controls;
   int char_narrowing;